requirements.txt
google-genai
python-dotenv
streamlit
langgraph
chromadb
sentence-transformers
pydantic